+++ src/main/java/jscover/filesystem/ConfigurationForFS.java
@@ -368,7 +368,7 @@ public class ConfigurationForFS extends Configuration {
     public static final String HELP_PREFIX2 = Main.HELP_PREFIX2;
     public static final String EXLCUDE_PREFIX = "--exclude=";
     public static final String EXLCUDE_REG_PREFIX = "--exclude-reg=";
-    public static final String INSTRUMENT_ONLY_REG_PREFIX = "--instrument-only-reg=";
+    public static final String ONLY_INSTRUMENT_REG_PREFIX = "--only-instrument-reg=";
     public static final String NO_INSTRUMENT_PREFIX = "--no-instrument=";
     public static final String NO_INSTRUMENT_REG_PREFIX = "--no-instrument-reg=";
     public static final String BRANCH_PREFIX = "--no-branch";
@@ -469,8 +469,8 @@ public class ConfigurationForFS extends Configuration {
                     configuration.showHelp = true;
                     configuration.invalid = true;
                 }
-            } else if (arg.startsWith(INSTRUMENT_ONLY_REG_PREFIX)) {
-                String patternString = arg.substring(INSTRUMENT_ONLY_REG_PREFIX.length());
+            } else if (arg.startsWith(ONLY_INSTRUMENT_REG_PREFIX)) {
+                String patternString = arg.substring(ONLY_INSTRUMENT_REG_PREFIX.length());
                 if (patternString.startsWith("/"))
                     patternString = patternString.substring(1);
                 configuration.defaultSkip = true;